moved WeekExample to a separate component + made Baseline externally controlled from top menu
